Vignette: General header style in vignette also format of function names writing style in some parts of the vignette I simplified the generation of the sorted list of genes for GSEA (no need to remove Inf)
Functions:
RES.filter is not a good idea: since running GSEA.run takes a long time, it doesn not make sense running it over a large number of datasets and the only retrieve part of the results. To reduce the number of datasets to analyze we already have the parameter Chip_index. The objective of RES.filter is to reduce the number of data to plot and hence it should be included in the plotting function. Thus, I modified plot_all_RES function to accept a list of TF and/or tables to plot. TODO: includeit in the otherplotfunctions as well? Remove RES.filter from GSEA.run?